moved token validation function to public validation module
