+++ src/lib.rs
@@ -356,6 +356,7 @@ pub mod provider;
 mod standard_claims;
 mod token;
 mod userinfo;
+pub mod validation;
 
 #[cfg(any(feature = "uma2", doc))]
 /// UMA2 OIDC/OAuth2 extension.
